refactor(attribution): migrate log_analyzer internals to svc subpackage
Move the service-layer implementation (analysis pipeline, job tracking,
splitlog, parsers, config, LLM output handling, etc.) from
`attribution/log_analyzer/` into `attribution/svc/`. The `log_analyzer`
package is reduced to a thin docstring shim pointing at the new location.

All internal imports across `analyzer/engine.py`, `analyzer/__init__.py`,
`attribution/__init__.py`, and `services/nvrx_attrsvc/service.py` are
updated to reference `attribution.svc.*` instead of
`attribution.log_analyzer.*`. Public re-exports from `attribution/__init__.py`
are preserved so the external API is unchanged.

Files deleted from log_analyzer/:
  analysis_pipeline.py, config.py, job.py, llm_output.py, log_analyzer.py,
  log_path_metadata.py, parser_base.py, posting_markdown.py, runner.py,
  slurm_parser.py, splitlog.py, tracked_jobs.py, types.py, utils.py
