+++ src/nvidia_resiliency_ext/attribution/svc/log_analyzer.py
@@ -18,25 +18,8 @@ import re
 from typing import Any, Awaitable, Callable, Dict, List, Optional, Tuple
 
 from nvidia_resiliency_ext.attribution.coalescing import LogAnalysisCoalesced
-from nvidia_resiliency_ext.attribution.log_analyzer.analysis_pipeline import (
-    AnalysisPipelineMode,
-    run_attribution_pipeline,
-)
-from nvidia_resiliency_ext.attribution.log_analyzer.config import ErrorCode, LogSageExecutionConfig
-from nvidia_resiliency_ext.attribution.log_analyzer.job import Job
-from nvidia_resiliency_ext.attribution.log_analyzer.log_path_metadata import CYCLE_LOG_PATTERN
 from nvidia_resiliency_ext.attribution.log_analyzer.nvrx_logsage import NVRxLogAnalyzer
-from nvidia_resiliency_ext.attribution.log_analyzer.splitlog import SplitlogTracker
-from nvidia_resiliency_ext.attribution.log_analyzer.tracked_jobs import TrackedJobs
-from nvidia_resiliency_ext.attribution.log_analyzer.types import (
-    LogAnalyzerError,
-    LogAnalyzerFilePreview,
-    LogAnalyzerSubmitResult,
-)
-from nvidia_resiliency_ext.attribution.log_analyzer.utils import (
-    nvrx_run_result_to_log_dict,
-    validate_log_path,
-)
+from .log_path_metadata import CYCLE_LOG_PATTERN
 from nvidia_resiliency_ext.attribution.mcp_integration import create_mcp_client
 from nvidia_resiliency_ext.attribution.path_utils import path_is_under_allowed_root
 from nvidia_resiliency_ext.attribution.postprocessing import post_analysis_items
@@ -46,6 +29,24 @@ from nvidia_resiliency_ext.attribution.trace_analyzer.fr_support import (
     fr_result_from_mcp_module_response,
 )
 from nvidia_resiliency_ext.attribution.trace_analyzer.trace_analyzer import TraceAnalyzer
+
+from .analysis_pipeline import (
+    AnalysisPipelineMode,
+    run_attribution_pipeline,
+)
+from .config import ErrorCode, LogSageExecutionConfig
+from .job import Job
+from .splitlog import SplitlogTracker
+from .tracked_jobs import TrackedJobs
+from .types import (
+    LogAnalyzerError,
+    LogAnalyzerFilePreview,
+    LogAnalyzerSubmitResult,
+)
+from .utils import (
+    nvrx_run_result_to_log_dict,
+    validate_log_path,
+)
 
 logger = logging.getLogger(__name__)
